requirements.txt
Django==1.9.2
docutils==0.12
python-slugify==1.2.0
